+++ spec/integration_spec.rb
@@ -234,7 +234,7 @@ describe_cli 'jazzy' do
       module_path = `swift build --build-path #{build_path} --show-bin-path`
       behaves_like cli_spec 'misc_jazzy_symgraph_features',
                             '--swift-build-tool symbolgraph ' \
-                              "--build-tool-arguments -I=#{module_path} "
+                              "--build-tool-arguments -I,#{module_path} "
     end
   end if !spec_subset || spec_subset == 'swift'